requirements.txt
pip==24.0
pytest==8.0.1
pandas==2.2.0
streamlit==1.31.1
plotly==5.18.0
clickhouse-driver==0.2.6
python-dotenv==1.0.1
dash-bootstrap-components==1.5.0